buypage: create test component
